# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/notes":{"post":{"operationId":"NotesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNoteDto"}}}},"responses":{"201":{"description":""}},"tags":["Notes"]},"get":{"operationId":"NotesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Notes"]}},"/notes/search":{"get":{"operationId":"NotesController_searchNote","parameters":[{"name":"q","required":false,"in":"query","description":"Search query string to filter notes by title or text.","example":"meeting notes","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notes"]}},"/notes/{id}":{"get":{"operationId":"NotesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notes"]},"put":{"operationId":"NotesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateNoteDto"}}}},"responses":{"200":{"description":""}},"tags":["Notes"]},"delete":{"operationId":"NotesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notes"]}}},"info":{"title":"Note App API","description":"The Note App API documentation","version":"1.0","contact":{}},"tags":[{"name":"notes","description":""}],"servers":[],"components":{"schemas":{"CreateNoteDto":{"type":"object","properties":{"title":{"type":"string","example":"My First Note"},"text":{"type":"string","example":"This is a description of my note."},"color":{"type":"string","example":"bg-yellow-400"}},"required":["title","text","color"]},"UpdateNoteDto":{"type":"object","properties":{}}}}}